Client library no longer a wizard (#1820)
