+++ google-cloud-tools-plugin/resources/META-INF/plugin.xml
@@ -507,10 +507,8 @@
       <reference ref="GoogleCloudTools.UploadSourceToGCP"/>
       <separator/>
       <action id="GoogleCloudTools.CloudStorage" class="com.google.cloud.tools.intellij.gcs.GcsToolWindowAction"/>
-      <!--
       <separator/>
-      <action id="GoogleCloudTools.AddCloudLibrariesWizard" class="com.google.cloud.tools.intellij.apis.AddCloudLibrariesWizardAction"/>
-      -->
+      <action id="GoogleCloudTools.AddCloudLibraries" class="com.google.cloud.tools.intellij.apis.AddCloudLibrariesAction"/>
       <separator/>
       <action id="GoogleCloudTools.Feedback" class="com.google.cloud.tools.intellij.CloudToolsFeedbackAction"/>
