+++ google-cloud-tools-plugin/resources/messages/CloudToolsBundle.properties
@@ -218,6 +218,7 @@ cloud.libraries.documentation.link=Documentation
 cloud.libraries.menu.action.text=Add Cloud libraries...
 cloud.libraries.menu.action.description=Select the Cloud libraries to download and include in your application
 cloud.libraries.module.selector.label=Module:
+cloud.libraries.ok.button.text=Add Client Library
 cloud.libraries.source.link=Source
 cloud.libraries.status.label=Status: {0}
 cloud.libraries.version.label=Version: {0}
